chore: vite chunk splitting, ci clippy gate, and clippy fixes
- Split vendor JS into separate cached chunks (react, router, zustand)
- Add clippy CI job with -D warnings to catch non-idiomatic Rust
- Fix all existing clippy warnings across instrument-core
- Rebuild wasm-pkg to reflect Rust changes
- Mark §6.3, §6.4, §6.6 as done in development plan

+++ src-core/instrument-core/src/json/yaml_to_json.rs
@@ -52,7 +52,10 @@ fn default_output() -> YamlToJsonOutput {
 }
 
 fn parse_error_line_column(err: &serde_yaml::Error) -> (Option<usize>, Option<usize>) {
-    (err.location().map(|l| l.line() as usize), err.location().map(|l| l.column() as usize))
+    (
+        err.location().map(|l| l.line()),
+        err.location().map(|l| l.column()),
+    )
 }
 
 /// Convert YAML input to formatted JSON.
